Elevate Labs Internship | Task No 8 | Day 8
This Power BI dashboard project was developed as part of Task 8 for the Elevate Labs Data Analyst Internship. The goal of the task was to design an interactive and insightful Sales Dashboard for a fictional company named Optimex, providing a complete analytical overview of sales performance, profitability, customer behavior, and product trends.